+++ inc/rng.hpp
@@ -9,10 +9,10 @@ typedef unsigned int uint;
 uint rng();
 uint randInt(uint max);
 int randInt(int min, int max);
-double randDouble();
-double randDouble(double max);
-double randDouble(double min, double max);
-double normPdf(double d);
-template<typename T> T selectObject(uint nObjects, std::pair<T, double> *collection);
+float randDouble();
+float randDouble(float max);
+float randDouble(float min, float max);
+float normPdf(float d);
+template<typename T> T selectObject(uint nObjects, std::pair<T, float> *collection);
 
 #endif
